scheduler now pushes full camera flow and kills it

+++ agent/python/dockerhelper.py
@@ -56,8 +56,9 @@ def runImage(image, cpu_shares, mem_limit, disk_limit, network, ports, environme
     containers[taskID] = container
 
 def killContainer(taskID):
-    container = containers[taskID]
-    container.kill()
+    if taskID in containers:
+        container = containers[taskID]
+        container.kill()
 
 def getContainerStatus(taskID):
     container = containers[taskID]
